Keep the principal shape next to the payload that carries it
Reusing `UserNameOnly` for this meant importing the user models, which reach the
database package, which cannot be imported from the models the query-service
payloads are built out of. That was solved by moving `PrincipalKind` and
`UserNameOnly` into a new leaf module and re-exporting them -- relocating types
other code imports, to add two fields to one payload.

The payload does not need the enum. It is a wire format, so the kind is a
constrained string and the shape lives beside the input it belongs to.
`PrincipalKind` goes back to being declared once, where it was.
